Add files via upload
Digest of individual files which are provided by brokerage Charles Schwab for stocks in various buckets; watchlists are the category of bucket!  For example item one, 3.15.2022 includes stocks that are followed since March 15, 2022.  This watchlist, for instance shows attributes of each stock including symbol, values for the price at c.o.b. (close-of-business), the price change (either positive or negative), the percentage move (calculated as the close minus previous day's closing price), and date.  For this repository the analysis includes a few watchlists summing to one hundred fifty (150 or so stocks.  The date for the analysis is the March 5, 2025.