Add files via upload
The file includes the code for map reduce work (hadoop) that contains the mapper, reducer code and the python container where it is written.  The mapper ingests data from the repository labeled as .csv (comma seperated values), data consumed by end users at brokerages (eg. Charles Schwab in this case).  The mapper parallelizes the query for data by field, example, stock ticker, delta of price; the mapper assigns the delta values to the key in key/value (k/v) pairs so the delta of price is stored as k/v with the value representing the stock symbol for that record, eg., aapl, 3% (shows 3 percent delta day over day for aapl on March 5 2024), the mapper assigns the ticker to the bin it falls under.  Bin 1 for example will store the record of all stock tickers with a delta of 1, eg., 1: spy, qqq, ibm, dell, cmcsa, aa, aal, bny, while another bin might read, 2: nvda, wmt, da, smh, ibm, boeing.  The reducer code is an aggregation of othe records of the apper.  So the bins will show the sum of tickers in each record; in this example 1 bin will show 8, 2 bin will show6.  The result of all two hundred stocks selected are produced.  For this study the result is shown in the "viewe_equities_Mar,5_2024.png".